requirements.txt
# Core
pyyaml>=6.0
cryptography>=41.0
pydantic>=2.5

# Async & Networking
aiohttp>=3.9
websockets>=12.0
httpx>=0.27

# Quantum Simulation (lightweight)
numpy>=1.26
scipy>=1.12

# AI/ML
scikit-learn>=1.4

# API
fastapi>=0.109
uvicorn>=0.27

# Testing
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=4.1

# Monitoring
prometheus-client>=0.20

# Utilities
structlog>=24.1
